% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-simu.R
\name{data.simu.null}
\alias{data.simu.null}
\title{Data simulation (null model & multiple SNPs)}
\usage{
data.simu.null(N, nSNP, nCov, maf, prev, TypeOfCov = NULL,
  CoefOfCov = NULL)
}
\arguments{
\item{N}{a numeric value: number of samples}

\item{nSNP}{a numeric value: number of SNPs}

\item{nCov}{a numeric value: number of covariates}

\item{maf}{a numeric value or a numeric vector of length nSNP: minor allele frequencies to simulate genotypes.
It should be between 0 and 0.5.}

\item{prev}{a numeric value: the expected proportion of cases among all samples.
It should be between 0 and 1.}

\item{TypeOfCov}{a character vector of length nCOV to specify the covariates types: "binary" ~ Bernoulli(0.5), "continuous" ~ Normal(0,1).
Default value is NULL, that is, 'rep(c("binary","continous"), length.out=nCov)'.}

\item{CoefOfCov}{a numeric vector of length nCOV to specify the covariates coefficients.
Default value is NULL, that is, 'rep(0.5, length.out=nCov)'.}
}
\value{
an R list including the following elements
\item{Phen.mtx}{an R matrix (N * nCov+1) of phenotype and covariates}
\item{Geno.mtx}{an R matrix (N * nSNP) of genotypes}
}
\description{
Simulate a dataset which includes genotypes, covariates and phenotypes.
}
\examples{
# Specify all arguments
N = 10000;
nSNP = 100
nCov = 2
maf = 0.05;  # OR maf = runif(nSNP, 0, 0.5)
prev = 0.01;
# Data simulation process
Data.ls = data.simu.null(N, nSNP, nCov, maf, prev)
Data.ls$Phen.mtx[1:10,]
Data.ls$Geno.mtx[1:10,1:10]
}
