% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SPAGE.one.SNP}
\alias{SPAGE.one.SNP}
\title{SaddlePoint Approximation implementation of GxE analysis (One-SNP-version)}
\usage{
SPAGE.one.SNP(g, obj.null, Envn.mtx, Cutoff = 2,
  impute.method = "none", missing.cutoff = 0.15, min.maf = 0,
  Firth.cutoff = 0, BetaG.cutoff = 0.001, BetaG.SPA = F)
}
\arguments{
\item{g}{a numeric genotype vector. Given that A(a) is the major(minor) allele, genotype should be coded as 0, 1, 2, and NA for AA, Aa, aa, and missing, respectively.}

\item{others}{the same as function SPAGE.}
}
\value{
the same as function SPAGE.
}
\description{
One-SNP-version SPAGE function. This function is to facilitate users that prefer reading and analyzing genotype line-by-line.
}
\examples{
Data.ls = data.simu.null(N = 1000, nSNP = 10, nCov = 2, maf = 0.3, prev = 0.1)
Phen.mtx = Data.ls$Phen.mtx
obj.null = SPAGE_Null_Model(y~Cov1+Cov2, data=Phen.mtx, out_type="D")
Envn.mtx = Data.ls$Phen.mtx[,"Cov1",drop=FALSE]
## Check help(SPAGE) to better understand the output?
SPAGE.one.SNP(Data.ls$Geno.mtx[,"rs1"], obj.null, Envn.mtx)
}
