% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SPAGE}
\alias{SPAGE}
\title{SaddlePoint Approximation implementation of GxE analysis}
\usage{
SPAGE(obj.null, Envn.mtx, Geno.mtx, Cutoff = 2, impute.method = "none",
  missing.cutoff = 0.15, min.maf = 0, Firth.cutoff = 0,
  BetaG.cutoff = 0.001, BetaG.SPA = F)
}
\arguments{
\item{obj.null}{output object of function SPAGE_Null_Model.}

\item{Envn.mtx}{a numeric environment matrix with each row as an individual and each column as an environmental factor. Column names of environmental factors are required.}

\item{Geno.mtx}{a numeric genotype matrix with each row as an individual and each column as a genetic variant.
Column names of genetic variations are required. Given that A(a) is the major(minor) allele.
Genotype should be coded as 0, 1, 2, and NA for AA, Aa, aa, and missing, respectively.}

\item{Cutoff}{a numeric value (Default: 2) to specify the standard deviation cutoff to be used.
If the test statistic lies within the standard deviation cutoff of the mean, its p value is calculated based on normal distribution approximation, otherwise, its p value is calculated based on saddlepoint approximation.}

\item{impute.method}{a character string (default= "none") to specify the method to impute missing genotypes.
"bestguess" imputes missing genotypes as most likely values (0,1,2), "random" imputes missing genotypes by generating binomial(2,p) random variables (p is the MAF), and "fixed" imputes missing genotypes by assigning the mean genotype values (2p).}

\item{missing.cutoff}{a numeric value (default=0.15) to specify the cutoff of the missing rates of SNPs.
Any SNP with missing rates higher than the cutoff will be excluded from the analysis.}

\item{min.maf}{a numeric value (default=0) to specify the cutoff of the minimal MAF. Any SNP with MAF < cutoff will be excluded from the analysis.}

\item{Firth.cutoff}{a numeric value (default=0, no Firth output) to specify the p-value cutoff for Firth's test.
Only when the SPA p-value less than the cutoff, Firth's test p-value is calculated.}

\item{BetaG.cutoff}{a numeric value (default=0.001) to specify the p-value cutoff for betaG estimation. See details for more information.}

\item{BetaG.SPA}{a logical value (default=F) to determine p.value.BetaG is calculated based on SPA (TRUE) or a normal distribution approximation (FALSE).}
}
\value{
an R matrix with the following columns
\item{MAF}{Minor allele frequencies}
\item{missing.rate}{Missing rate}
\item{p.value.BetaG}{p value of the marginal genotypic effect based on normal distribution approximation (BetaG.SPA=F) or saddlepoint approximation (BetaG.SPA=T)}
\item{p.value.spa-xx}{p value of the marginal GxE effect based on saddlepoint approximation. xx is the name of the environmental factor}
\item{p.value.norm-xx}{p value of the marginal GxE effect based on the normal distribution approximation. xx is the name of the environmental factor}
\item{p.value.Firth-xx}{p value of the marginal GxE effect based on the Firth's test. xx is the name of the environmental factor. xx is the name of the environmental factor}
\item{Stat-xx}{test statistic of the marginal GxE effect. xx is the name of the environmental factor}
\item{Var-xx}{estimated variance of the marginal GxE effect. xx is the name of the environmental factor}
\item{z-xx}{z score of the marginal GxE effect. xx is the name of the environmental factor}
}
\description{
Test for association between marginal GxE multiplicative interaction effect and dichotomous phenotypes.
}
\details{
Here we propose a scalable and accurate method, SPAGE (SaddlePoint Approximation implementation of G×E analysis), that is applicable for genome-wide scale phenome-wide G×E studies (PheWIS).
SPAGE fits a genotype-independent logistic model only once across the whole-genome analysis to reduce computation cost and uses a saddlepoint approximation (SPA) to calibrate the test statistics for analysis of phenotypes with unbalanced case-control ratios.
When genotypic effect is small or moderate (true for most of the variants), the method can control type I error rates well.
We first test for the marginal genotypic effect (normal approximation if Beta.SPA=F and SPA if Beta.SPA=T) and if the p value is less than the pre-given argument 'BetaG.cutoff', we will update the test statistic and p value.
}
\examples{
# Specify all arguments
Data.ls = data.simu.null(N = 1000, nSNP = 10, nCov = 2, maf = 0.3, prev = 0.1)
Phen.mtx = Data.ls$Phen.mtx
obj.null = SPAGE_Null_Model(y~Cov1+Cov2, data=Phen.mtx, out_type="D")
Envn.mtx = as.matrix(Phen.mtx)[,"Cov1",drop=FALSE]
SPAGE(obj.null, Envn.mtx, Data.ls$Geno.mtx)
}
