% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-null.R
\name{SPAGE_Null_Model}
\alias{SPAGE_Null_Model}
\title{Get parameters and residuals from the NULL model}
\usage{
SPAGE_Null_Model(formula, data = NULL, out_type = "D")
}
\arguments{
\item{formula}{an object of class “formula”: a symbolic description of the Covariates-Only model to be fitted.}

\item{data}{an optional data frame containing the variables in the model (default=NULL). If it is NULL, the variables are taken from 'environment(formula)'}

\item{out_type}{an indicator of the outcome type. "D" for the dichotomous outcome.}
}
\value{
an R object to pass to function SPAGE for GxE analyses.
}
\description{
Fit a covariates-only model to compute model parameters and residuals for SPAGE.
}
\examples{
Data.ls = data.simu.null(N = 1000, nSNP = 10, nCov = 2, maf = 0.3, prev = 0.1)
Phen.mtx = Data.ls$Phen.mtx
obj.null = SPAGE_Null_Model(y~Cov1+Cov2, data=Phen.mtx, out_type="D")
}
